requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
blinker==1.4
boto3==1.24.0
botocore==1.27.0; python_version >= '3.7'
certifi==2022.5.18.1; python_version >= '3.6'
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3; python_version >= '3.7'
cryptography==36.0.2
datadog==0.44.0
flask-seasurf==1.1.1
flask-talisman==1.0.0
flask==2.1.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3; python_version >= '3'
importlib-metadata==4.11.4; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.0; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
marshmallow-dataclass==8.5.8
marshmallow-jsonschema==0.13.0
marshmallow==3.14.1
mypy-extensions==0.4.3
neo4j==4.4.3
neobolt==1.7.17
pdpyras==4.5.0
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.4.0
pynamodb==5.2.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-json-logger==2.0.2
pytz==2022.1
pyyaml==6.0
requests==2.27.1
s3transfer==0.6.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slack-sdk==3.17.0
typing-extensions==4.2.0; python_version >= '3.7'
typing-inspect==0.7.1
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.1.2; python_version >= '3.7'
zipp==3.8.0; python_version >= '3.7'
zope.event==4.5.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
